Raports. Log in with SMS. Delayed orders.
Generation of a raport about required orders with ability to print them. (List of orders for one day or few selected days)

Now admin is able to log in to the Admin Panel only using SMS-verification. Also set Admin Session for PC equal 8 hours and session for Mobile Phone endless.

Added functionality of detection of the delayed orders with counting of penalty bill.